Move MerkleDatabase to libsawtooth
Also moves StateReader trait, StateIter and StateDatabaseError.

This commit also ignores a doc test that was not being run before
but now fails.
