+++ libsawtooth/src/state/error.rs
@@ -20,7 +20,8 @@ use std::fmt;
 use cbor::decoder::DecodeError;
 use cbor::encoder::EncodeError;
 use protobuf::ProtobufError;
-use sawtooth::database::error::DatabaseError;
+
+use crate::database::error::DatabaseError;
 
 #[derive(Debug)]
 pub enum StateDatabaseError {
