+++ validator/Cargo.toml
@@ -10,14 +10,10 @@ hex = "0.3"
 lazy_static = "1.0"
 log = { version = "0.4", features = ["std"] }
 libc = ">=0.2.35"
-cbor-codec = "0.7"
 protobuf = "2.0"
 python3-sys = "0.2"
 sawtooth = { path = "../libsawtooth", features = ["validator-internals"] }
 uluru = "0.2.0"
-
-[dev-dependencies]
-rand = "0.4"
 
 [build-dependencies]
 protoc-rust = "2.0"
