+++ validator/src/lib.rs
@@ -15,7 +15,6 @@
  * ------------------------------------------------------------------------------
  */
 
-extern crate cbor;
 extern crate cpython;
 extern crate hex;
 extern crate libc;
@@ -25,8 +24,6 @@ extern crate python3_sys as py_ffi;
 extern crate lazy_static;
 #[macro_use]
 extern crate log;
-#[cfg(test)]
-extern crate rand;
 extern crate sawtooth;
 extern crate uluru;
