+++ validator/src/state/state_pruning_manager.rs
@@ -18,9 +18,9 @@ use std::cmp::Ordering;
 use std::collections::BinaryHeap;
 
 use sawtooth::database::lmdb::LmdbDatabase;
+use sawtooth::state::merkle::MerkleDatabase;
 
 use metrics;
-use state::merkle::MerkleDatabase;
 
 lazy_static! {
     static ref COLLECTOR: metrics::MetricsCollectorHandle =
