+++ validator/src/state/state_view_factory.rs
@@ -15,10 +15,9 @@
  * ------------------------------------------------------------------------------
  */
 use sawtooth::database::lmdb::LmdbDatabase;
-
-use state::error::StateDatabaseError;
-use state::merkle::{DecodedMerkleStateReader, MerkleDatabase};
-use state::StateReader;
+use sawtooth::state::error::StateDatabaseError;
+use sawtooth::state::merkle::{DecodedMerkleStateReader, MerkleDatabase};
+use sawtooth::state::StateReader;
 
 /// The StateViewFactory produces StateViews for a particular merkle root.
 ///
